+++ internal/checks/rule_for.go
@@ -16,8 +16,16 @@ const (
 	RuleForCheckName = "rule/for"
 )
 
-func NewRuleForCheck(minFor, maxFor time.Duration, severity Severity) RuleForCheck {
+type RuleForKey string
+
+const (
+	RuleForFor           RuleForKey = "for"
+	RuleForKeepFiringFor RuleForKey = "keep_firing_for"
+)
+
+func NewRuleForCheck(key RuleForKey, minFor, maxFor time.Duration, severity Severity) RuleForCheck {
 	return RuleForCheck{
+		key:      key,
 		minFor:   minFor,
 		maxFor:   maxFor,
 		severity: severity,
@@ -26,6 +34,7 @@ func NewRuleForCheck(minFor, maxFor time.Duration, severity Severity) RuleForChe
 
 type RuleForCheck struct {
 	severity Severity
+	key      RuleForKey
 	minFor   time.Duration
 	maxFor   time.Duration
 }
@@ -50,11 +59,22 @@ func (c RuleForCheck) Check(_ context.Context, _ string, rule parser.Rule, _ []d
 	var forDur model.Duration
 	var fragment string
 	var lines []int
-	if rule.AlertingRule.For != nil {
-		forDur, _ = model.ParseDuration(rule.AlertingRule.For.Value.Value)
-		fragment = rule.AlertingRule.For.Value.Value
-		lines = rule.AlertingRule.For.Lines()
+
+	switch c.key {
+	case RuleForFor:
+		if rule.AlertingRule.For != nil {
+			forDur, _ = model.ParseDuration(rule.AlertingRule.For.Value.Value)
+			fragment = rule.AlertingRule.For.Value.Value
+			lines = rule.AlertingRule.For.Lines()
+		}
+	case RuleForKeepFiringFor:
+		if rule.AlertingRule.KeepFiringFor != nil {
+			forDur, _ = model.ParseDuration(rule.AlertingRule.KeepFiringFor.Value.Value)
+			fragment = rule.AlertingRule.KeepFiringFor.Value.Value
+			lines = rule.AlertingRule.KeepFiringFor.Lines()
+		}
 	}
+
 	if fragment == "" {
 		fragment = rule.AlertingRule.Alert.Value.Value
 		lines = rule.AlertingRule.Alert.Lines()
@@ -65,7 +85,7 @@ func (c RuleForCheck) Check(_ context.Context, _ string, rule parser.Rule, _ []d
 			Fragment: fragment,
 			Lines:    lines,
 			Reporter: c.Reporter(),
-			Text:     fmt.Sprintf("this alert rule must have a 'for' field with a minimum duration of %s", output.HumanizeDuration(c.minFor)),
+			Text:     fmt.Sprintf("this alert rule must have a '%s' field with a minimum duration of %s", c.key, output.HumanizeDuration(c.minFor)),
 			Severity: c.severity,
 		})
 	}
@@ -75,7 +95,7 @@ func (c RuleForCheck) Check(_ context.Context, _ string, rule parser.Rule, _ []d
 			Fragment: fragment,
 			Lines:    lines,
 			Reporter: c.Reporter(),
-			Text:     fmt.Sprintf("this alert rule must have a 'for' field with a maximum duration of %s", output.HumanizeDuration(c.maxFor)),
+			Text:     fmt.Sprintf("this alert rule must have a '%s' field with a maximum duration of %s", c.key, output.HumanizeDuration(c.maxFor)),
 			Severity: c.severity,
 		})
 	}